% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvintDAG.R
\name{rmvDAG_int}
\alias{rmvDAG_int}
\title{A random graph data generation function}
\usage{
rmvDAG_int(n, amat, bmat, Sig = NULL)
}
\arguments{
\item{n}{The sample size}

\item{amat}{The adjacency matrix of a DAG}

\item{bmat}{The intervention matrix}

\item{Sig}{The error variance of each node}
}
\value{
Gaussian data with the given sample size
}
\description{
This function generates random observations from a DAG graph
}
\examples{
amat <- matrix(c(0,1,0,0),2,2)
bmat <- matrix(c(1,1,0,1),2,2)
rmvDAG_int(50, amat, bmat)
}
