% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvDAG.R
\name{rmvDAG_obs}
\alias{rmvDAG_obs}
\title{A random graph data generation function}
\usage{
rmvDAG_obs(n, amat, Sig = NULL)
}
\arguments{
\item{n}{The sample size}

\item{amat}{The adjacency matrix of a DAG}

\item{Sig}{The error variance of each node}
}
\value{
Gaussian data with the given sample size
}
\description{
This function generates random observations from a DAG graph
}
\examples{
amat=matrix(c(0,1,0,0),2,2)
rmvDAG_obs(50,amat)
}
